backend/requirements.txt
wheel==0.45.1
setuptools==80.9.0
typing-extensions==4.15.0


Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv>=1.1.1
whitenoise>=6.6.0
gunicorn>=23.0.0

pymongo>=4.5
psycopg==3.2.10  # for Postgres if used


google-genai>=1.36.0
google-generativeai>=0.8.5
httpx>=0.28.1
pydantic>=2.11.7
tenacity>=9.1.2


requests>=2.32.5
numpy>=2.3.3
pillow>=11.3.0 
